pyproject.toml
[project]
name = "electric-meter"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "bokeh==3.4.0",
    "pandas==2.2.1",
    "requests==2.31.0",
    "python-dotenv==1.0.1",
]

[project.scripts]
dataviz = "electric_meter.serve:main"
fetch-daily-data = "electric_meter.fetch_daily_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/electric_meter"]
